Replace Atom references from README.md
Atom is a [deprecated app](https://github.blog/2022-06-08-sunsetting-atom/), so using it as an example of installable app through `brew` could be annoying for some users. I replaced it with a similar app name (VS Code).
Also, fixed the `alt` in the git, as it is not showing an Atom installation, but Alfred.